004: made class HighlightButton separately
